Fix tests, provide h2 Liquidbase support, and update documentation
