+++ owasp-dependency-check-suppressions.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<suppressions xmlns="https://jeremylong.github.io/DependencyCheck/dependency-suppression.1.3.xsd">
+    <!-- Add your suppression rules here, or leave empty for now -->
+</suppressions>
