+++ src/main/java/edu/tamu/scholars/middleware/view/model/CollectionView.java
@@ -28,7 +28,7 @@ public abstract class CollectionView extends View {
     private Layout layout;
 
     @ElementCollection
-    @MapKeyColumn(name = "key")
+    @MapKeyColumn(name = "\"key\"")
     @Column(name = "template", columnDefinition = "TEXT")
     private Map<String, String> templates;
