+++ src/site/markdown/index.md
@@ -1,0 +1,44 @@
+# VIVO Scholars Middleware
+
+Welcome to the VIVO Scholars Discovery Middleware documentation site.
+
+## Overview
+
+VIVO Scholars Discovery Middleware is a Spring Boot application for managing scholarly individual, directory, discovery and data and analytic views.
+
+## Quick Links
+
+- [API Documentation](apidocs/index.html) - JavaDoc API documentation
+- [REST API Documentation](docs/index.html) - REST endpoint documentation  
+- [Code Coverage Report](jacoco/index.html) - JaCoCo test coverage analysis
+- [Security Report](security/dependency-check-report.html) - OWASP dependency vulnerability analysis
+
+## Project Information
+
+- **Organization:** Texas A&M University Libraries
+- **License:** MIT License
+- **Source Code:** [GitHub Repository](https://github.com/TAMULib/scholars-discovery)
+
+## Getting Started
+
+This middleware application provides:
+
+- Scholarly data management and discovery
+- RESTful APIs for data access
+- Integration with VIVO ontologies
+- Advanced search and filtering capabilities
+- Data analytics and reporting features
+
+## Build and Reports
+
+This site contains comprehensive project documentation including:
+
+- Project dependencies and plugin information
+- Code quality metrics via Checkstyle
+- Test coverage analysis via JaCoCo
+- Security vulnerability assessments via OWASP Dependency Check
+- Complete API documentation
+
+## Support
+
+For questions and support, please visit our [GitHub repository](https://github.com/TAMULib/scholars-discovery) or contact the Texas A&M University Libraries Development Team.
