+++ src/test/java/edu/tamu/scholars/middleware/auth/config/WebSecurityConfigTest.java
@@ -6,11 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ContextConfiguration;
 
-import edu.tamu.scholars.middleware.config.SolrTestConfig;
-
-@ContextConfiguration(classes = SolrTestConfig.class)
 // @formatter:off
 @SpringBootTest(properties = {
     "spring.profiles.active=default",
