+++ src/test/java/edu/tamu/scholars/middleware/auth/controller/LoginControllerTest.java
@@ -6,7 +6,7 @@ import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
-import static org.springframework.restdocs.request.RequestDocumentation.queryParameters;
+import static org.springframework.restdocs.request.RequestDocumentation.formParameters;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -48,7 +48,7 @@ public class LoginControllerTest extends UserIntegrationTest {
                 .andDo(
                     document(
                         "login",
-                        queryParameters(
+                        formParameters(
                             describeUser.withParameter("username", "The username to login as, usually an e-mail address."),
                             describeUser.withParameter("password", "The password associated with the specified username.")
                         ),
