+++ src/test/resources/logback-test.xml
@@ -5,6 +5,7 @@
   <logger name="org.apache" level="ERROR"/>
   <logger name="org.docx4j" level="ERROR"/>
   <logger name="org.hibernate" level="OFF"/>
+  <logger name="liquibase" level="OFF"/>
   <logger name="org.springframework" level="ERROR"/>
   <logger name="edu.tamu.scholars.middleware" level="WARN"/>
 </configuration>
